+++ frontend/src/core/components/ProfileSidebar/ProfileSidebar.scss
