+++ frontend/src/views/Profile/ProfileLayout/ProfileLayout.scss
